[기능개선] BPM 변화가 클 때 딜레이 추가
[구조개선] 객체 렌더방식 롤백
- 오브젝트 풀 ActivateObjects 구조의 LinkedList에 삽입 삭제 시 가비지 생성으로 인한 롤백 및 LinkedList 제거

+++ Assets/Scripts/Global/GameSetting.cs
@@ -54,8 +54,8 @@ public class GameSetting
     public static PitchType      CurrentPitchType     = PitchType.None;
 
     // Speed
-    private static double OriginScrollSpeed = 6.5d; 
-    public static double ScrollSpeed
+    private static float OriginScrollSpeed = 6.5f; 
+    public static float ScrollSpeed
     {
 
         get => OriginScrollSpeed;
@@ -66,8 +66,8 @@ public class GameSetting
         }
     }
 
-    public static double Weight => ScrollSpeed * 350d;
-    public static double PreLoadTime => 1200d / Weight;
+    public static float Weight => ScrollSpeed * 350f;
+    public static float PreLoadTime => 1200f / Weight;
 
     // Sound
     public static readonly float DefaultSoundOffset = -50;
